// ui/src/openapi/models/ApiAssignment.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ShpanKids API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ApiAssignmentStatus } from './ApiAssignmentStatus';
import {
    ApiAssignmentStatusFromJSON,
    ApiAssignmentStatusFromJSONTyped,
    ApiAssignmentStatusToJSON,
} from './ApiAssignmentStatus';
import type { ApiAssignmentType } from './ApiAssignmentType';
import {
    ApiAssignmentTypeFromJSON,
    ApiAssignmentTypeFromJSONTyped,
    ApiAssignmentTypeToJSON,
} from './ApiAssignmentType';

/**
 * 
 * @export
 * @interface ApiAssignment
 */
export interface ApiAssignment {
    /**
     * 
     * @type {string}
     * @memberof ApiAssignment
     */
    id: string;
    /**
     * 
     * @type {ApiAssignmentType}
     * @memberof ApiAssignment
     */
    type: ApiAssignmentType;
    /**
     * 
     * @type {string}
     * @memberof ApiAssignment
     */
    title: string;
    /**
     * 
     * @type {Date}
     * @memberof ApiAssignment
     */
    forDate: Date;
    /**
     * 
     * @type {ApiAssignmentStatus}
     * @memberof ApiAssignment
     */
    status: ApiAssignmentStatus;
    /**
     * 
     * @type {string}
     * @memberof ApiAssignment
     */
    description?: string;
}

/**
 * Check if a given object implements the ApiAssignment interface.
 */
export function instanceOfApiAssignment(value: object): boolean {
    if (!('id' in value)) return false;
    if (!('type' in value)) return false;
    if (!('title' in value)) return false;
    if (!('forDate' in value)) return false;
    if (!('status' in value)) return false;
    return true;
}

export function ApiAssignmentFromJSON(json: any): ApiAssignment {
    return ApiAssignmentFromJSONTyped(json, false);
}

export function ApiAssignmentFromJSONTyped(json: any, ignoreDiscriminator: boolean): ApiAssignment {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'type': ApiAssignmentTypeFromJSON(json['type']),
        'title': json['title'],
        'forDate': (new Date(json['forDate'])),
        'status': ApiAssignmentStatusFromJSON(json['status']),
        'description': json['description'] == null ? undefined : json['description'],
    };
}

export function ApiAssignmentToJSON(value?: ApiAssignment | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'id': value['id'],
        'type': ApiAssignmentTypeToJSON(value['type']),
        'title': value['title'],
        'forDate': ((value['forDate']).toISOString()),
        'status': ApiAssignmentStatusToJSON(value['status']),
        'description': value['description'],
    };
}